+++ src/main/kotlin/dev/aaronhowser/mods/aaron/item/ItemWithTooltip.kt
@@ -1,0 +1,24 @@
+package dev.aaronhowser.mods.aaron.item
+
+import net.minecraft.network.chat.Component
+import net.minecraft.world.item.Item
+import net.minecraft.world.item.ItemStack
+import net.minecraft.world.item.TooltipFlag
+
+class ItemWithTooltip(
+	properties: Properties,
+	private val tooltipProvider: (ItemStack) -> List<Component>
+) : Item(properties) {
+
+	constructor(properties: Properties, tooltips: List<Component>) : this(properties, { tooltips })
+
+	override fun appendHoverText(
+		stack: ItemStack,
+		context: TooltipContext,
+		tooltipComponents: MutableList<Component>,
+		tooltipFlag: TooltipFlag
+	) {
+		tooltipComponents.addAll(tooltipProvider.invoke(stack))
+	}
+
+}