+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - A LOT of backend renderer changes
 - Advancement stuff
 - MutableList#shuffle
+- AaronLootTableDsl
 
 # 1.16.0